# api-docs.yaml
openapi: "3.0.0"
info:
  title: Gear & Grit - Gear Rental API
  description: |
    API documentation for the Gear & Grit platform - a gear rental marketplace.
    Users can register as CUSTOMERs or PROVIDERs. Customers can browse gear, create rental orders, and make payments.
    Providers can manage their gear inventory and fulfill orders. Admins can manage users and view all system data.
  version: "1.0.0"
  contact:
    name: API Support
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token. Can be provided as Bearer token in Authorization header, or as an httpOnly cookie named 'accessToken'."
    CookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
      description: "JWT access token stored in httpOnly cookie. Also accepts refreshToken cookie for token refresh."
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        statusCode:
          type: integer
          example: 200
        message:
          type: string
          example: "Operation successful"
        data:
          nullable: true
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        errorDetails:
          nullable: true
          type: string

    # ============================================
    # Auth Schemas
    # ============================================
    RegisterInput:
      type: object
      required:
        - name
        - email
        - password
        - role
      properties:
        name:
          type: string
          description: User's full name
          example: "John Doe"
        email:
          type: string
          format: email
          description: User's email address
          example: "john@example.com"
        password:
          type: string
          minLength: 6
          description: Password (minimum 6 characters)
          example: "password123"
        role:
          type: string
          enum: [CUSTOMER, PROVIDER]
          description: User role (ADMIN registration is not allowed)
          example: "CUSTOMER"
    LoginInput:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "john@example.com"
        password:
          type: string
          example: "password123"
    UpdateProfileInput:
      type: object
      properties:
        name:
          type: string
          description: Updated name
          example: "John Updated"
        avatarUrl:
          type: string
          format: uri
          description: Updated avatar URL
          example: "https://example.com/avatar.jpg"
    AuthUserResponse:
      type: object
      properties:
        id:
          type: string
          example: "clx..."
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          example: "john@example.com"
        role:
          type: string
          example: "CUSTOMER"
        avatarUrl:
          type: string
          nullable: true
    LoginResponseData:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AuthUserResponse"

    # ============================================
    # Category Schemas
    # ============================================
...
  responses:
...
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized - not logged in or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
...
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: |
        Register a new user account as CUSTOMER or PROVIDER.
        ADMIN role cannot be registered through this endpoint.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AuthUserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: |
        Authenticate a user with email and password.
        Sets httpOnly cookies for accessToken and refreshToken.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: User logged in successfully
          headers:
            Set-Cookie:
              schema:
                type: string
                description: "accessToken and refreshToken cookies set"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/LoginResponseData"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/auth/refresh-token:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: |
        Refresh the access token using the refreshToken cookie.
        Requires valid refreshToken cookie to be present.
      operationId: refreshToken
      responses:
        "200":
          description: Token refreshed successfully
          headers:
            Set-Cookie:
              schema:
                type: string
                description: "New accessToken cookie set"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          accessToken:
                            type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /api/auth/update-profile:
    put:
      tags:
        - Authentication
      summary: Update user profile
      description: Update the name and/or avatar URL of the currently authenticated user.
      operationId: updateProfile
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileInput"
      responses:
        "200":
          description: Profile updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AuthUserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"

  # ============================================
  # Category Routes
  # ============================================
...